Move browser folder add edit component to vault
