+++ apps/browser/src/vault/popup/settings/folder-add-edit.component.html
